+++ src/constants.js
@@ -14,6 +14,12 @@ export const itemStatuses = {
   DECLARED_LOST: 'Declared lost',
 };
 
+export const loanActions = {
+  CLAIMED_RETURNED: 'claimReturned',
+  DECLARE_LOST: 'declareLost',
+  MARK_AS_MISSING: 'markAsMissing',
+};
+
 export const deliveryFulfillmentValues = {
   HOLD_SHELF: 'Hold Shelf',
   DELIVERY: 'Delivery',
@@ -28,3 +34,10 @@ export const statusFilter = [
   { label: 'ui-users.filters.status.active', value: 'active' },
   { label: 'ui-users.filters.status.inactive', value: 'inactive' }
 ];
+
+/* With current id determines that this is fee/fine condition,
+ because this condition is validating with different message,
+ and condition fields are rendering dinamically.
+ All conditions (thare are 6 of them) are always present on BE
+ with hardcoded ids for now. */
+export const feeFineBalanceId = 'cf7a0d5f-a327-4ca1-aa9e-dc55ec006b8a';