+++ src/views/UserSearch/UserSearch.js
@@ -60,6 +60,9 @@ class UserSearch extends React.Component {
     }).isRequired,
     mutator: PropTypes.shape({
       loans: PropTypes.object,
+      resultOffset: PropTypes.shape({
+        replace: PropTypes.func.isRequired,
+      }),
     }).isRequired,
     source: PropTypes.object,
     visibleColumns: PropTypes.arrayOf(PropTypes.string),
@@ -277,6 +280,7 @@ class UserSearch extends React.Component {
       onNeedMoreData,
       resources,
       contentRef,
+      mutator: { resultOffset },
     } = this.props;
 
     const users = get(resources, 'records.records', []);
@@ -402,8 +406,9 @@ class UserSearch extends React.Component {
                             </div>
                             <Filters
                               activeFilters={activeFilters.state}
-                              resources={this.props.resources}
+                              resources={resources}
                               onChangeHandlers={getFilterHandlers()}
+                              resultOffset={resultOffset}
                             />
                           </form>
                         </Pane>
@@ -444,6 +449,8 @@ class UserSearch extends React.Component {
                           autosize
                           virtualize
                           hasMargin
+                          pageAmount={100}
+                          pagingType="click"
                         />
 
                       </Pane>